gh-18 changed name of files with theme
