+++ lib/core/theme/i_theme/i_theme.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 part 'i_profile_theme.dart';
 part 'i_login_theme.dart';
 part 'i_dish_card_theme.dart';
-part 'i_registration_card.dart';
+part 'i_registration_card_theme.dart';
 part 'i_data_input_card_theme.dart';
 
 abstract class ITheme {
