+++ lib/core/theme/light_theme/light_theme.dart
@@ -5,7 +5,7 @@ part 'light_profile_theme.dart';
 part 'light_login_theme.dart';
 part 'light_dish_card_theme.dart';
 part 'light_data_input_card_theme.dart';
-part 'light_registration_card.dart';
+part 'light_registration_card_theme.dart';
 
 const Color _yellow = Color.fromRGBO(251, 187, 0, 1);
 const Color _orange = Color.fromRGBO(248, 113, 34, 1);
@@ -20,7 +20,7 @@ class LightTheme implements ITheme {
   static LightLoginTheme _loginTheme = LightLoginTheme();
   static LightDishCardTheme _dishCardTheme = LightDishCardTheme();
   static LightDataInputCardTheme _dataInputCardTheme =
-  LightDataInputCardTheme();
+      LightDataInputCardTheme();
   static LightRegistrationCardTheme _registrationCardTheme =
       LightRegistrationCardTheme();
